A Maven POM for Cassandra.
This POM:
* uses dependencies only from public maven repos.
* compiles the classes.
* attempts to run some of the unit tests (those in test/unit).
* respects assertions.

This POM does not:
* generate the thrift interface files.
* generate the cql lexer.
* run all the unit tests properly.

Properly running the unit tests requires the test harness to do some housekeeping external to the tests. See the `testmacrohelper` macrodef to understand what those things are.

The script needs to be augmented to run the tests located in test/burn, test/long, and test/microbench with their proper environmental settings.

Baby steps.